InT/requirements.txt
altair==5.3.0
api==0.0.7
asgiref==3.6.0
astroid==2.15.5
attrs==21.2.0
Automat==20.2.0
Babel==2.8.0
bcrypt==4.0.1
blinker==1.4
boto3==1.34.117
botocore==1.34.117
cachetools==5.3.3
certifi==2023.5.7
cffi==1.16.0
chardet==4.0.0
charset-normalizer==3.1.0
click==8.0.3
cloud-init==24.1.3
colorama==0.4.4
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==42.0.7
DateTime==5.5
dbus-python==1.2.18
defusedxml==0.7.1
dill==0.3.8
distro==1.7.0
distro-info==1.1+ubuntu0.2
dj-rest-auth==5.0.1
Django==4.0.3
django-allauth==0.57.0
django-cors-headers==4.0.0
django-environ==0.11.2
django-ses==4.1.0
django-webpack-loader==1.8.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
ec2-hibinit-agent==1.0.0
environ==1.0
git-remote-codecommit==1.17
gitdb==4.0.11
GitPython==3.1.43
hibagent==1.0.1
httplib2==0.20.2
hyperlink==21.0.0
idna==3.4
ikp3db==1.4.2
importlib-metadata==4.6.4
incremental==21.3.0
isort==5.12.0
jedi==0.19.0
jeepney==0.7.1
Jinja2==3.0.3
jmespath==1.0.1
jsonpatch==1.32
jsonpointer==2.0
jsonschema==3.2.0
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy-object-proxy==1.10.0
markdown-it-py==3.0.0
MarkupSafe==2.0.1
mccabe==0.7.0
mdurl==0.1.2
more-itertools==8.10.0
mysqlclient==2.2.0
netifaces==0.11.0
nose==1.3.7
numpy==1.26.4
oauthlib==3.2.2
packaging==24.0
pandas==2.2.2
parso==0.8.4
pexpect==4.8.0
pillow==10.3.0
platformdirs==4.2.2
protobuf==4.25.3
ptyprocess==0.7.0
pyarrow==16.1.0
pyasn1==0.4.8
pyasn1-modules==0.2.1
pycparser==2.21
pydeck==0.9.1
Pygments==2.18.0
PyGObject==3.42.1
PyHamcrest==2.0.2
PyJWT==2.8.0
pylint==2.17.4
pylint-django==2.5.3
pylint-flask==0.6
pylint-plugin-utils==0.8.2
PyMySQL==1.1.0
pyOpenSSL==24.1.0
pyparsing==2.4.7
pyrsistent==0.18.1
pyserial==3.5
python-apt==2.4.0+ubuntu3
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.1
python-magic==0.4.24
python3-openid==3.2.0
pytz==2023.3
PyYAML==5.4.1
requests==2.30.0
requests-oauthlib==1.3.1
rich==13.7.1
s3transfer==0.10.1
SecretStorage==3.3.1
service-identity==18.1.0
six==1.16.0
smmap==5.0.1
sos==4.5.6
sqlparse==0.4.3
ssh-import-id==5.11
streamlit==1.35.0
streamlit-chat==0.1.1
systemd-python==234
tenacity==8.3.0
toml==0.10.2
tomli==2.0.1
tomlkit==0.12.5
toolz==0.12.1
tornado==6.4.1
Twisted==22.1.0
typing_extensions==4.12.0
tzdata==2022.7
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==2.0.2
wadllib==1.3.6
watchdog==4.0.1
wrapt==1.16.0
zipp==1.0.0
zope.interface==5.4.0
